+++ merlin/models/tf/outputs/sampling/in_batch.py
