+++ tests/unit/tf/outputs/test_dot_product.py
